removed duplicate mgcpmediagroup class #1614
