+++ restcomm/restcomm.mrb/pom.xml
@@ -90,6 +90,13 @@
 
 		<dependency>
 			<groupId>org.restcomm</groupId>
+			<artifactId>restcomm-connect.mscontrol.mms</artifactId>
+			<version>${project.version}</version>
+			<scope>provided</scope>
+		</dependency>
+
+		<dependency>
+			<groupId>org.restcomm</groupId>
 			<artifactId>restcomm-connect.mrb.api</artifactId>
 			<version>${project.version}</version>
 			<scope>provided</scope>
